fix(connector): update Aliyun SMS connector config
