+++ packages/connectors/connector-aliyun-sms/src/index.test.ts
@@ -1,6 +1,5 @@
 import { TemplateType } from '@logto/connector-kit';
 
-import { isChinaNumber } from './index.js';
 import { mockedConnectorConfig, phoneTest, codeTest } from './mock.js';
 
 const getConfig = vi.fn().mockResolvedValue(mockedConnectorConfig);
@@ -15,36 +14,6 @@ vi.mock('./single-send-text.js', () => ({
 }));
 
 const { default: createConnector } = await import('./index.js');
-
-describe('isChinaNumber()', () => {
-  it('should validate China phone numbers in non-strict mode', () => {
-    // Valid cases
-    expect(isChinaNumber('13812345678')).toBe(true);
-    expect(isChinaNumber('8613812345678')).toBe(true);
-    expect(isChinaNumber('008613812345678')).toBe(true);
-    expect(isChinaNumber('+8613812345678')).toBe(true);
-
-    // Invalid cases
-    expect(isChinaNumber('1381234567')).toBe(false); // Too short
-    expect(isChinaNumber('138123456789')).toBe(false); // Too long
-    expect(isChinaNumber('abcdefghijk')).toBe(false); // Non-numeric
-    expect(isChinaNumber('8513812345678')).toBe(false); // Wrong prefix
-  });
-
-  it('should validate China phone numbers in strict mode', () => {
-    // Valid cases
-    expect(isChinaNumber('8613812345678', true)).toBe(true);
-    expect(isChinaNumber('008613812345678', true)).toBe(true);
-    expect(isChinaNumber('+8613812345678', true)).toBe(true);
-
-    // Invalid cases
-    expect(isChinaNumber('13812345678', true)).toBe(false); // Missing region code
-    expect(isChinaNumber('1381234567', true)).toBe(false); // Too short
-    expect(isChinaNumber('138123456789', true)).toBe(false); // Too long
-    expect(isChinaNumber('abcdefghijk', true)).toBe(false); // Non-numeric
-    expect(isChinaNumber('8513812345678', true)).toBe(false); // Wrong prefix
-  });
-});
 
 describe('sendMessage()', () => {
   afterEach(() => {
