+++ packages/connectors/connector-aliyun-sms/src/utils.test.ts
@@ -1,12 +1,11 @@
 import { mockedParameters } from './mock.js';
 
 const post = vi.fn();
-
 vi.mock('got', () => ({
   got: { post },
 }));
 
-const { getSignature, request } = await import('./utils.js');
+const { getSignature, request, isChinaNumber } = await import('./utils.js');
 
 describe('getSignature', () => {
   it('should get valid signature', () => {
@@ -31,5 +30,35 @@ describe('request', () => {
     expect(payload.Timestamp).not.toBeNull();
     expect(payload.SignatureNonce).not.toBeNull();
     expect(payload.Signature).not.toBeNull();
+  });
+});
+
+describe('isChinaNumber()', () => {
+  it('should validate China phone numbers in non-strict mode', () => {
+    // Valid cases
+    expect(isChinaNumber('13812345678')).toBe(true);
+    expect(isChinaNumber('8613812345678')).toBe(true);
+    expect(isChinaNumber('008613812345678')).toBe(true);
+    expect(isChinaNumber('+8613812345678')).toBe(true);
+
+    // Invalid cases
+    expect(isChinaNumber('1381234567')).toBe(false); // Too short
+    expect(isChinaNumber('138123456789')).toBe(false); // Too long
+    expect(isChinaNumber('abcdefghijk')).toBe(false); // Non-numeric
+    expect(isChinaNumber('8513812345678')).toBe(false); // Wrong prefix
+  });
+
+  it('should validate China phone numbers in strict mode', () => {
+    // Valid cases
+    expect(isChinaNumber('8613812345678', true)).toBe(true);
+    expect(isChinaNumber('008613812345678', true)).toBe(true);
+    expect(isChinaNumber('+8613812345678', true)).toBe(true);
+
+    // Invalid cases
+    expect(isChinaNumber('13812345678', true)).toBe(false); // Missing region code
+    expect(isChinaNumber('1381234567', true)).toBe(false); // Too short
+    expect(isChinaNumber('138123456789', true)).toBe(false); // Too long
+    expect(isChinaNumber('abcdefghijk', true)).toBe(false); // Non-numeric
+    expect(isChinaNumber('8513812345678', true)).toBe(false); // Wrong prefix
   });
 });
