+++ packages/connectors/connector-aliyun-sms/src/utils.ts
@@ -62,3 +62,12 @@ export const request = async (
     form: { ...finalParameters, Signature: signature },
   });
 };
+
+// When `isStrictOnRegionNumber` is true, it means the mobile phone number area code needs to be strictly matched; the default is false, for backward compatibility, as some old phone numbers do not have region codes, which would cause the method to make mistakes on validation.
+export const isChinaNumber = (to: string, isStrictOnRegionNumber?: boolean) => {
+  if (isStrictOnRegionNumber) {
+    return /^(\+86|0086|86)\d{11}$/.test(to);
+  }
+
+  return /^(\+86|0086|86)?\d{11}$/.test(to);
+};
